server/rss_fetcher/requirements.txt
Authlib==1.3.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.4
feedparser==6.0.11
idna==3.6
numpy==1.26.4
pandas==2.2.0
pycparser==2.21
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.5
tzdata==2024.1
urllib3==2.2.1
validators==0.22.0
weaviate-client==4.4.4
